chore: move files from 'super roast bot' to 'super-roast-bot' so changes apply to actual codebase
